refactored some drawing stuff to systems
