+++ imgui.ini
@@ -15,6 +15,6 @@ Collapsed=0
 
 [Window][Object Info]
 Pos=0,30
-Size=640,833
+Size=768,1013
 Collapsed=0
